# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ThesisCreate": {
        "additionalProperties": false,
        "properties": {
          "anti_sectors": {
            "items": {
              "type": "string"
            },
            "title": "Anti Sectors",
            "type": "array"
          },
          "check_size_max": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Size Max"
          },
          "check_size_min": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Size Min"
          },
          "currency": {
            "default": "USD",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          },
          "deal_breakers": {
            "items": {
              "type": "string"
            },
            "title": "Deal Breakers",
            "type": "array"
          },
          "geos": {
            "items": {
              "type": "string"
            },
            "title": "Geos",
            "type": "array"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "must_haves": {
            "items": {
              "type": "string"
            },
            "title": "Must Haves",
            "type": "array"
          },
          "name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "ownership_target_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ownership Target Pct"
          },
          "risk_appetite": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "0.5",
            "title": "Risk Appetite"
          },
          "sectors": {
            "items": {
              "type": "string"
            },
            "title": "Sectors",
            "type": "array"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Weights",
            "type": "object"
          }
        },
        "required": [
          "name"
        ],
        "title": "ThesisCreate",
        "type": "object"
      },
      "ThesisOut": {
        "properties": {
          "anti_sectors": {
            "items": {
              "type": "string"
            },
            "title": "Anti Sectors",
            "type": "array"
          },
          "check_size_max": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Size Max"
          },
          "check_size_min": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Size Min"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "deal_breakers": {
            "items": {
              "type": "string"
            },
            "title": "Deal Breakers",
            "type": "array"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "geos": {
            "items": {
              "type": "string"
            },
            "title": "Geos",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "must_haves": {
            "items": {
              "type": "string"
            },
            "title": "Must Haves",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "ownership_target_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ownership Target Pct"
          },
          "risk_appetite": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Risk Appetite",
            "type": "string"
          },
          "sectors": {
            "items": {
              "type": "string"
            },
            "title": "Sectors",
            "type": "array"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Weights",
            "type": "object"
          }
        },
        "required": [
          "id",
          "org_id",
          "name",
          "is_default",
          "sectors",
          "anti_sectors",
          "stages",
          "geos",
          "check_size_min",
          "check_size_max",
          "currency",
          "ownership_target_pct",
          "risk_appetite",
          "weights",
          "must_haves",
          "deal_breakers",
          "notes",
          "version",
          "deleted_at",
          "created_at",
          "updated_at"
        ],
        "title": "ThesisOut",
        "type": "object"
      },
      "ThesisPreview": {
        "properties": {
          "excluded_by_gate": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Excluded By Gate",
            "type": "object"
          },
          "matched_count": {
            "title": "Matched Count",
            "type": "integer"
          },
          "sample": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sample",
            "type": "array"
          }
        },
        "required": [
          "matched_count",
          "sample",
          "excluded_by_gate"
        ],
        "title": "ThesisPreview",
        "type": "object"
      },
      "ThesisUpdate": {
        "additionalProperties": false,
        "properties": {
          "anti_sectors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Sectors"
          },
          "check_size_max": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Size Max"
          },
          "check_size_min": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Size Min"
          },
          "currency": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deal_breakers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deal Breakers"
          },
          "geos": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geos"
          },
          "must_haves": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Must Haves"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "ownership_target_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ownership Target Pct"
          },
          "risk_appetite": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Appetite"
          },
          "sectors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sectors"
          },
          "stages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stages"
          },
          "weights": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights"
          }
        },
        "title": "ThesisUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Evidence-backed venture sourcing, screening, diligence, and decisions",
    "title": "VC Brain API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/theses": {
      "get": {
        "operationId": "list_theses_api_v1_theses_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThesisOut"
                  },
                  "title": "Response List Theses Api V1 Theses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Theses",
        "tags": [
          "theses"
        ]
      },
      "post": {
        "operationId": "create_thesis_api_v1_theses_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThesisCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Thesis",
        "tags": [
          "theses"
        ]
      }
    },
    "/api/v1/theses/{thesis_id}": {
      "delete": {
        "operationId": "delete_thesis_api_v1_theses__thesis_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "thesis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thesis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Thesis",
        "tags": [
          "theses"
        ]
      },
      "get": {
        "operationId": "get_thesis_api_v1_theses__thesis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "thesis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thesis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Thesis",
        "tags": [
          "theses"
        ]
      },
      "patch": {
        "operationId": "update_thesis_api_v1_theses__thesis_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "thesis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thesis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThesisUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Thesis",
        "tags": [
          "theses"
        ]
      }
    },
    "/api/v1/theses/{thesis_id}/default": {
      "post": {
        "operationId": "set_default_api_v1_theses__thesis_id__default_post",
        "parameters": [
          {
            "in": "path",
            "name": "thesis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thesis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Default",
        "tags": [
          "theses"
        ]
      }
    },
    "/api/v1/theses/{thesis_id}/preview": {
      "post": {
        "operationId": "preview_thesis_api_v1_theses__thesis_id__preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "thesis_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thesis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Thesis",
        "tags": [
          "theses"
        ]
      }
    },
...
  }
}